Extract class to own file
